DEV-428 exclude SearchInput from Form
