+++ src/assets/css/otp-styles-bundle.less
@@ -48,7 +48,7 @@
 @import '../../components/typography/ctypography';
 @import '../../components/divider/cdivider';
 @import '../../components/_anchorCard/canchorcard';
-@import '../../components/form/SearchInput/csearch-input';
+@import '../../components/searchinput/csearch-input';
 @import "../../components/carousel/ccarousel";
 
 // Responsive utils
