--- src/components/form/SearchInput/index.less
@@ -1,2 +1,0 @@
-@import '../../../assets/css/bundle.less';
-@import 'csearch-input.less';
