+++ src/components/searchinput/index.less
@@ -1,0 +1,2 @@
+@import '../../assets/css/bundle.less';
+@import 'csearch-input.less';
