+++ src/index.tsx
@@ -29,3 +29,4 @@ export * from './components/pageheader';
 export * from './components/divider';
 export * from './components/_anchorCard';
 export * from './components/carousel';
+export * from './components/searchinput';
